LayoutTests/accessibility/auto-fill-types.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<input type="text" value="hello" id="textfield">

<script>
    let output = "This tests that the auto-filled buttons show up.\n";

    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        let textField = document.getElementById("textfield");
        let axTextField = accessibilityController.accessibleElementById("textfield");
        output += `Initial auto-fill available: ${axTextField.boolAttributeValue("AXValueAutofillAvailable")}\n`;
        output += `Auto-fill type: ${axTextField.stringAttributeValue("AXValueAutofillType")}\n`;

        window.internals.setShowAutoFillButton(document.getElementById("textfield"), "Contacts");
        setTimeout(async function() {
            let contactsButton = null;
            // Wait for contacts autofill button to render.
            await waitFor(() => {
                contactsButton = axTextField.childAtIndex(1);
                return contactsButton && axTextField.stringAttributeValue("AXValueAutofillType") === "contacts";
            });
            output += `Contact button role: ${contactsButton.role}\n`;
            output += `Contact button label: ${contactsButton.description}\n`;
            output += `Auto-fill type: ${axTextField.stringAttributeValue("AXValueAutofillType")}\n`;

            window.internals.setShowAutoFillButton(document.getElementById("textfield"), "Credentials");
            // Wait for credentials autofill button to render.
            let credentialsButton = null;
            await waitFor(() => {
                credentialsButton = axTextField.childAtIndex(1);
                return credentialsButton && axTextField.stringAttributeValue("AXValueAutofillType") === "credentials"
                    && credentialsButton.description.includes("password");
            });
            output += `Credentials button role: ${credentialsButton.role}\n`;
            output += `Credentials button label: ${credentialsButton.description}\n`;
            output += `Auto-fill type: ${axTextField.stringAttributeValue("AXValueAutofillType")}\n`;

            output += `Post auto-fill available: ${axTextField.boolAttributeValue("AXValueAutofillAvailable")}\n`;

            debug(output);
            finishJSTest();
        }, 0);
    }
</script>
</body>
</html>
