LayoutTests/accessibility/auto-fill-types-expected.txt

This tests that the auto-filled buttons show up.
Initial auto-fill available: false
Auto-fill type: none
Contact button role: AXRole: AXButton
Contact button label: AXDescription: contact info AutoFill
Auto-fill type: contacts
Credentials button role: AXRole: AXButton
Credentials button label: AXDescription: password AutoFill
Auto-fill type: credentials
Post auto-fill available: true

PASS successfullyParsed is true

TEST COMPLETE
